refator: recoil state 파일 분리
