+++ src/screens/CreateTransactionStack/AccountCategoryScreen.tsx
@@ -1,7 +1,7 @@
 import {useMemo} from 'react';
 import {useRecoilState} from 'recoil';
 import {isEmpty} from 'lodash';
-import {accountState} from 'libs/recoil/states/transaction';
+import {accountState} from 'libs/recoil/states/account';
 import {CreateTransactionStackRouteProp} from 'navigators/types';
 import ScreenContainer from 'components/@common/ScreenContainer';
 import LeftButton from 'components/CreateTransactionStack/LeftButton';
